+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/person/update/UpdatePersonView.groovy
@@ -1,9 +1,16 @@
 package life.qbic.portal.offermanager.components.person.update
 
-
+import com.vaadin.data.provider.ListDataProvider
+import com.vaadin.icons.VaadinIcons
+import com.vaadin.shared.ui.grid.HeightMode
+import com.vaadin.ui.Button
+import com.vaadin.ui.Grid
+import com.vaadin.ui.Label
+import com.vaadin.ui.components.grid.HeaderRow
 import groovy.util.logging.Log4j2
 import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.portal.offermanager.components.AppViewModel
+import life.qbic.portal.offermanager.components.GridUtils
 import life.qbic.portal.offermanager.components.person.create.CreatePersonController
 import life.qbic.portal.offermanager.components.person.create.CreatePersonView
 
@@ -20,6 +27,8 @@ class UpdatePersonView extends CreatePersonView{
     private final UpdatePersonViewModel updatePersonViewModel
     private final AppViewModel sharedViewModel
 
+    private Grid<Affiliation> affiliations
+    private Button addAffiliationButton
 
     UpdatePersonView(CreatePersonController controller, AppViewModel sharedViewModel, UpdatePersonViewModel updatePersonViewModel) {
         super(controller, sharedViewModel, updatePersonViewModel)
@@ -32,6 +41,89 @@ class UpdatePersonView extends CreatePersonView{
     private void adjustViewElements() {
         submitButton.caption = "Update Person"
         abortButton.caption = "Abort Person Update"
+
+        //add a grid
+        affiliations = new Grid<>()
+        generateAffiliationGrid()
+        affiliations.setCaption("Current Affiliations")
+        affiliations.setSelectionMode(Grid.SelectionMode.NONE)
+
+        this.addComponent(affiliations,2)
+        //add a heading for adding a new affiliation
+        Label newAffiliation = new Label("Add a new affiliation")
+        this.addComponent(newAffiliation,3)
+
+        //add the add button
+        addAffiliationButton = new Button("Add Affiliation")
+        addAffiliationButton.setIcon(VaadinIcons.PLUS)
+        addAffiliationButton.setEnabled(false)
+
+        buttonLayout.addComponent(addAffiliationButton,0)
+    }
+
+    private void generateAffiliationGrid() {
+        try {
+            this.affiliations.addColumn({ affiliation -> affiliation.category.value })
+                    .setCaption("Category").setId("Category")
+            this.affiliations.addColumn({ affiliation -> affiliation.organisation })
+                    .setCaption("Organisation").setId("Organisation")
+            this.affiliations.addColumn({ affiliation -> affiliation.addressAddition })
+                    .setCaption("Address Addition").setId("AddressAddition")
+            this.affiliations.addColumn({ affiliation -> affiliation.street })
+                    .setCaption("Street").setId("Street")
+            this.affiliations.addColumn({ affiliation -> affiliation.city })
+                    .setCaption("City").setId("City")
+            this.affiliations.addColumn({ affiliation -> affiliation.postalCode })
+                    .setCaption("Postal Code").setId("PostalCode")
+            this.affiliations.addColumn({ affiliation -> affiliation.country })
+                    .setCaption("Country").setId("Country")
+
+            //specify size of grid and layout
+            affiliations.setWidthFull()
+            affiliations.setHeightMode(HeightMode.ROW)
+            affiliations.setHeightByRows(5)
+
+        } catch (Exception e) {
+            new Exception("Unexpected exception in building the affiliation grid", e)
+        }
+
+        /*
+        Let's not forget to setup the grid's data provider
+        */
+        def affiliationDataProvider = setupAffiliationDataProvider()
+        /*
+        Lastly, we add some content filters for the columns
+         */
+        addFilters(affiliationDataProvider)
+    }
+
+    private ListDataProvider setupAffiliationDataProvider() {
+        def affiliationListDataProvider = new ListDataProvider<>(updatePersonViewModel.affiliationList)
+        this.affiliations.setDataProvider(affiliationListDataProvider)
+
+        return affiliationListDataProvider
+    }
+
+    private void addFilters(ListDataProvider affiliationListDataProvider) {
+        HeaderRow personFilterRow = affiliations.appendHeaderRow()
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("Organisation"),
+                personFilterRow)
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("AddressAddition"),
+                personFilterRow)
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("Street"),
+                personFilterRow)
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("City"),
+                personFilterRow)
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("PostalCode"),
+                personFilterRow)
+        GridUtils.setupColumnFilter(affiliationListDataProvider,
+                affiliations.getColumn("Country"),
+                personFilterRow)
     }
 
     private void registerListener(){
@@ -43,11 +135,9 @@ class UpdatePersonView extends CreatePersonView{
                 String firstName = updatePersonViewModel.firstName
                 String lastName = updatePersonViewModel.lastName
                 String email = updatePersonViewModel.email
-                List<Affiliation> affiliations = new ArrayList()
-                affiliations.add(updatePersonViewModel.affiliation)
+                List<Affiliation> affiliations = updatePersonViewModel.affiliationList
 
                 if(updatePersonViewModel.outdatedPerson){
-                    affiliations.addAll(updatePersonViewModel.outdatedPerson.affiliations)
                     controller.updatePerson(updatePersonViewModel.outdatedPerson, firstName, lastName, title, email, affiliations)
                 }
 
@@ -60,5 +150,72 @@ class UpdatePersonView extends CreatePersonView{
                 sharedViewModel.failureNotifications.add("An unexpected error occurred. We apologize for any inconveniences. Please inform us via email to support@qbic.zendesk.com.")
             }
         })
+
+        updatePersonViewModel.addPropertyChangeListener("affiliationValid",{
+            if(updatePersonViewModel.affiliationValid){
+                addAffiliationButton.setEnabled(true)
+            }else{
+                addAffiliationButton.setEnabled(false)
+            }
+        })
+
+        addAffiliationButton.addClickListener({
+            if(!updatePersonViewModel.affiliationList.contains(updatePersonViewModel.affiliation)){
+                updatePersonViewModel.affiliationList << updatePersonViewModel.affiliation
+                affiliations.dataProvider.refreshAll()
+                updatePersonViewModel.personUpdated = true
+            }else{
+                sharedViewModel.failureNotifications.add("Cannot add the selected affiliation. It was already associated with the person.")
+                addAffiliationButton.setEnabled(false)
+            }
+            resetAffiliation()
+        })
+
+        updatePersonViewModel.addPropertyChangeListener({it ->
+            if(updatePersonViewModel.outdatedPerson){
+                switch (it.propertyName) {
+                    case "academicTitle":
+                        boolean titleChanged = updatePersonViewModel.academicTitle != updatePersonViewModel.outdatedPerson.title.toString()
+                        updatePersonViewModel.personUpdated = updatePersonViewModel.academicTitleValid && titleChanged
+                        break
+                    case "firstName":
+                        boolean firstNameChanged = updatePersonViewModel.firstName != updatePersonViewModel.outdatedPerson.firstName
+                        updatePersonViewModel.personUpdated = updatePersonViewModel.firstNameValid && firstNameChanged
+                        break
+                    case "lastName":
+                        boolean lastNameChanged = updatePersonViewModel.lastName != updatePersonViewModel.outdatedPerson.lastName
+                        updatePersonViewModel.personUpdated = updatePersonViewModel.lastNameValid && lastNameChanged
+                        break
+                    case "email":
+                        boolean emailChanged = updatePersonViewModel.email != updatePersonViewModel.outdatedPerson.emailAddress
+                        updatePersonViewModel.personUpdated = updatePersonViewModel.emailValid && emailChanged
+                        break
+                    default:
+                        break
+                }
+                submitButton.enabled = allValuesValid()
+            }
+        })
+    }
+
+    private void resetAffiliation(){
+        organisationComboBox.selectedItem = organisationComboBox.clear()
+        addressAdditionComboBox.selectedItem = addressAdditionComboBox.clear()
+        addressAdditionComboBox.setComponentError(null)
+    }
+
+    /**
+     * This is used to indicate whether all fields of this view are filled correctly.
+     * It relies on the separate fields for validation. It overrides the Method allValuesValid() of the CreatePersonView
+     * and makes the validity independent of an added affiliation. Furthermore, the person entry needs to be different to
+     * the person that needs to be updated.
+     *
+     * @return boolean which indicates if a person update can be triggered
+     */
+    protected boolean allValuesValid() {
+        return createPersonViewModel.firstNameValid \
+            && createPersonViewModel.lastNameValid \
+            && createPersonViewModel.emailValid
+            && updatePersonViewModel.personUpdated
     }
 }