+++ offer-manager-app/src/main/groovy/life/qbic/portal/offermanager/components/person/update/UpdatePersonViewModel.groovy
@@ -1,5 +1,7 @@
 package life.qbic.portal.offermanager.components.person.update
 
+import groovy.beans.Bindable
+import life.qbic.datamodel.dtos.business.Affiliation
 import life.qbic.datamodel.dtos.business.Customer
 import life.qbic.datamodel.dtos.general.Person
 
@@ -28,6 +30,9 @@ import life.qbic.portal.offermanager.dataresources.persons.ProjectManagerResourc
 class UpdatePersonViewModel extends CreatePersonViewModel{
 
     final private EventEmitter<Person> customerUpdate
+    ObservableList affiliationList
+
+    @Bindable Boolean personUpdated
 
     UpdatePersonViewModel(CustomerResourceService customerService,
             ProjectManagerResourceService managerResourceService,
@@ -36,10 +41,12 @@ class UpdatePersonViewModel extends CreatePersonViewModel{
             PersonResourceService personResourceService) {
         super(customerService, managerResourceService, affiliationService, personResourceService)
         this.customerUpdate = customerUpdate
+        affiliationList = new ArrayList<Affiliation>()
+        personUpdated = false
 
         this.customerUpdate.register((Person person) -> {
-            loadData(person)
             setOutdatedPerson(person)
+            loadData(person)
         })
     }
 
@@ -48,6 +55,8 @@ class UpdatePersonViewModel extends CreatePersonViewModel{
         firstName = person.firstName
         lastName = person.lastName
         email = person.emailAddress
-        affiliation = person.affiliations.first()
+        //obtain the affiliations
+        affiliationList.clear()
+        affiliationList.addAll(person.affiliations)
     }
 }